Verify that the new active controller is the same

+++ metadata/src/main/java/org/apache/kafka/controller/QuorumController.java
@@ -1373,7 +1373,7 @@ public final class QuorumController implements Controller {
     }
 
     @Override
-    public long curClaimEpoch() {
+    public int curClaimEpoch() {
         return curClaimEpoch;
     }
